[6172]- remove webpacker, add new gems
