+++ app/assets/config/manifest.js
@@ -1,1 +1,2 @@
 // blazer requires sprockets and this file.
+//= link_tree ../builds
