+++ app/assets/stylesheets/application.scss
@@ -1,6 +1,6 @@
 @import "./govuk_style_setup";
-@import "~govuk-frontend/govuk/all";
-@import "~@ministryofjustice/frontend/moj/all";
+@import "govuk-frontend/govuk/all";
+@import "@ministryofjustice/frontend/moj/all";
 @import "./govuk_overrides";
 @import "./filter_toggle_button";
 @import "./trainee_list";
