+++ app/assets/stylesheets/cookie_banner.scss
