+++ app/assets/stylesheets/filter_toggle_button.scss
