+++ app/assets/stylesheets/govuk_style_setup.scss
@@ -1,9 +1,9 @@
 @function frontend-font-url($filename) {
-  @return url("~assets/fonts/" + $filename);
+  @return url("govuk-frontend/govuk/assets/fonts/" + $filename);
 }
 
 @function frontend-image-url($filename) {
-  @return url("~assets/images/" + $filename);
+  @return url("govuk-frontend/govuk/assets/images/" + $filename);
 }
 
 $govuk-font-url-function: frontend-font-url;
