+++ app/assets/stylesheets/inverted.scss
