+++ app/assets/stylesheets/tables.scss
