+++ app/assets/stylesheets/trainee.scss
