+++ app/assets/stylesheets/trainee_list.scss
