+++ app/javascript/scripts/autocomplete/calculateWeight.spec.js
