+++ app/javascript/scripts/autocomplete/clean.spec.js
