+++ app/javascript/scripts/components.js
@@ -1,0 +1,5 @@
+function importAll (r) {
+  r.keys().forEach(r)
+}
+
+importAll(require.context('../../../components', true, /script.js$/))
