+++ app/javascript/scripts/cookie_banner.js
