+++ app/javascript/scripts/cookie_banner.spec.js
