+++ app/javascript/scripts/filter_toggle.js
