+++ app/javascript/scripts/live_filter.js
