+++ app/javascript/scripts/live_filter.spec.js
