+++ app/javascript/scripts/tracker.spec.js
