+++ app/javascript/scripts/utils/cookie_helper.js
