+++ app/views/layouts/application.html.erb
@@ -24,6 +24,7 @@
       'https://www.googletagmanager.com/gtm.js?id='+i+dl+ '&gtm_auth=<%= Settings.google_tag_manager.auth_id %>&gtm_preview=env-<%= Settings.google_tag_manager.env_id %>&gtm_cookies_win=x';f.parentNode.insertBefore(j,f);
       })(window,document,'script','dataLayer','<%= Settings.google_tag_manager.tracking_id %>');</script>
     <% end %>
+    <%= javascript_include_tag "application", "data-turbo-track": "reload", defer: true %>
   </head>
 
   <body class="govuk-template__body <%= yield :body_class %>">
