--- app/webpacker/scripts/components.js
@@ -1,5 +1,0 @@
-function importAll (r) {
-  r.keys().forEach(r)
-}
-
-importAll(require.context('../../components', true, /script.js$/))
